+++ eva/src/components/dashBoardInside/dashMultiLine.vue
@@ -193,11 +193,11 @@ export default {
           this.props.nodata = false; // то убираем соощение о отсутствии данных
           this.props.result = this.dataRestFrom;  // заносим все данные в переменную
 
-          if (this.props.legends.length == 0) {
+          if (this.props.legends.length === 0) {
             this.metrics = [];
 
             let metricsName = Object.keys(this.props.result[0]).filter( item => {
-              if (item.indexOf('caption') == -1 && item.indexOf('annotation') == -1) {
+              if (item.indexOf('caption') === -1 && item.indexOf('annotation') === -1) {
                 return item
               }
             });
@@ -254,7 +254,7 @@ export default {
 
       let graphics = d3.select(this.$el.querySelector('.dash-multi')).selectAll('svg').nodes(); // получаем область в которой будем рисовтаь график
 
-      if(graphics.length != 0){  // если график уже есть
+      if(graphics.length !== 0){  // если график уже есть
         graphics[0].remove(); // удаляем его
       }
 
@@ -272,7 +272,7 @@ export default {
       let metricsName = [...[],...that.metrics];  // массив из всех метрик что доступны на графике и в нужном нам порядке
       metricsName.splice(0,1);
 
-      if(metricsName.length == 0){ // если метрик вообще не найдено
+      if(metricsName.length === 0){ // если метрик вообще не найдено
         this.props.nodata = true;  // показываем сообщение о некорректности данных
         this.props.result = [];  // очищаем массив результатов
         this.props.message = "Ни одной метрики не найдено. Проверьте корректность данных";  // выводим сообщение
@@ -349,11 +349,10 @@ export default {
       }
 
       let annotation = Object.keys(data[0]).filter( item => {
-        if (item.indexOf('annotation') != -1) {
+        if (item.indexOf('annotation') !== -1) {
           return item
         }
       })
-
 
       if (time) {
         !timeFormat ? timeFormat = '%Y-%m-%d %H:%M:%S' : false;
@@ -364,7 +363,7 @@ export default {
             .tickFormat(d3.timeFormat(timeFormat))
             .tickValues(
               x.ticks().filter( (item,i) => {
-                if (i%deliter == 0) {
+                if (i%deliter === 0) {
                   return item
                 }
               })
@@ -385,7 +384,7 @@ export default {
       verticalLineX();
 
       // создаем область графика, все-что вне этой области не будет отрисованно
-      let clip = svg.append("defs").append("svg:clipPath")
+      svg.append("defs").append("svg:clipPath")
         .attr("id", `clip-${that.id}`)
         .append("svg:rect")
         .attr("width", width )
@@ -481,17 +480,17 @@ export default {
           let allDotHover = [];
           if (extraDot.length > 0) {
             extraDot.forEach( (item,j) => {
-              if (metricsName[i] == item.column) {
+              if (metricsName[i] === item.column) {
                 nullValue = j
               }
             })
           }
 
-          if(nullValue == -1) {
+          if(nullValue === -1) {
 
             data.forEach( line => {
-              if (!Number(line[metricsName[i]]) && line[metricsName[i]] != 0) {
-                onelinesWithBreak.length == 1 ? mustSee.push(onelinesWithBreak[0]) : false;
+              if (!Number(line[metricsName[i]]) && line[metricsName[i]] !== 0) {
+                onelinesWithBreak.length === 1 ? mustSee.push(onelinesWithBreak[0]) : false;
                 linesWithBreak.push(onelinesWithBreak);
                 onelinesWithBreak = [];
               } else {
@@ -499,7 +498,7 @@ export default {
               }
             })
 
-            onelinesWithBreak.length == 1 ? mustSee.push(onelinesWithBreak[0]) : false;
+            onelinesWithBreak.length === 1 ? mustSee.push(onelinesWithBreak[0]) : false;
             linesWithBreak.push(onelinesWithBreak);
             AllLinesWithBreak.push(linesWithBreak);
             linesWithBreak.forEach( (lineItself,j) => {
@@ -532,7 +531,7 @@ export default {
             .attr("fill",colorLine[i])
             .style("opacity",function(d,j) {
               let opacity = 0;
-              if (nullValue != -1) {
+              if (nullValue !== -1) {
                 opacity =  1
               }
               mustSee.forEach( item => {
@@ -541,7 +540,7 @@ export default {
                 }
               })
               if (lastDot) {
-                if (j == data.length-1) { // если это последняя точка, то
+                if (j === data.length-1) { // если это последняя точка, то
                   opacity = 1;  // и постоянно ее отображаем
                   putLabelDot('data-last-dote',`last-dot-text-${metricsName[i]}`,d,y(d[metricsName[i]])-5,metricsName[i],this,that,'line',brushObj);
                 }
@@ -551,7 +550,7 @@ export default {
                 putLabelDot('data-with-caption',`caption-dot-text-${metricsName[i]}`,d,y(d[metricsName[i]])-5,`_${metricsName[i]}_caption`,this,that,'line',brushObj);
               }
 
-              if (annotation.length != 0) {
+              if (annotation.length !== 0) {
                 annotation.forEach( (item,i) => {
                   if (d[item]) {
                     verticalLine(d,item,metricsName.length+1+i,tooltip, this.strokewidth);
@@ -571,10 +570,10 @@ export default {
               }
               let text =  '';
               Object.keys(d).forEach( key => {
-                if (key == xMetric) {
+                if (key === xMetric) {
                   text += `<p><span>${key}</span> : ${xVal}</p>`;
                 } else {
-                  if (key.indexOf('annotation') == -1) {
+                  if (key.indexOf('annotation') === -1) {
                     text += `<p><span>${key}</span> : ${d[key]}</p>`;
                   }
 
@@ -604,7 +603,7 @@ export default {
 
               allDotHover = svg.selectAll('circle').nodes().filter( dot => {
                 if(dot.classList.contains('dot')){
-                  if (dot.getAttribute('cx') == x(d[xMetric]*secondTransf) && dot['__data__'][dot.getAttribute('metric')] != null ) {
+                  if (dot.getAttribute('cx') === x(d[xMetric]*secondTransf) && dot['__data__'][dot.getAttribute('metric')] != null ) {
                     dot.style="opacity:1";
                     return dot
                   }
@@ -620,13 +619,13 @@ export default {
             .on("mouseout", function(d) {
               let opacity = 1;
               if (!this.getAttribute("data-last-dote")) {
-                if (nullValue == -1) {
+                if (nullValue === -1) {
                   opacity = 0;
                 }
 
                 allDotHover.forEach( dot => {
                   //console.log(dot['__data__'])
-                  if (extraDot.indexOf(dot['__data__']) == -1) {
+                  if (extraDot.indexOf(dot['__data__']) === -1) {
                     dot.style=`opacity:0`;
                   }
                   if (dot.getAttribute("data-with-caption")) {
@@ -640,7 +639,7 @@ export default {
               }
 
               mustSee.forEach( item => {
-                if (item[metricsName[i]] == d[metricsName[i]]) {
+                if (item[metricsName[i]] === d[metricsName[i]]) {
                   opacity =  1
                 }
               })
@@ -707,14 +706,12 @@ export default {
               brushObj.selectionMove();
             })
             .on("mouseup", () => {
-
               brushObj.selectionUp();
             })
         }
 
         brushObj['selectionMove'] = () => {
           if (brushObj.mouseDown) {
-
             if ((event.offsetX-50 - brushObj.startX) > 0) {
               brushObj.direction = 'right';
               brushObj.endX = event.offsetX-50;
@@ -727,7 +724,6 @@ export default {
                 .attr("x", brushObj.startX + (event.offsetX-50 - brushObj.startX))
                 .attr("width", -(event.offsetX-50 - brushObj.startX))
             }
-
           }
         }
 
@@ -742,7 +738,6 @@ export default {
           if (brush.select(`.selection`).attr("width") > 5) {
             updateData([brushObj.startX,brushObj.endX],brushObj,-1)
           }
-
         }
 
         brushObj['clearBrush'] = () => {
@@ -755,8 +750,6 @@ export default {
         }
 
       } else {
-
-
         let step = ((height-20)/metricsName.length).toFixed(5);
         let startY = [20];
         let maxYTop,minYBottom;
@@ -792,7 +785,7 @@ export default {
 
         // создаем вертикальные линии
 
-        if (annotation.length != 0) {
+        if (annotation.length !== 0) {
           annotation.forEach( (item,i) => {
             data.forEach( d => {
               if (d[item]) {
@@ -864,7 +857,7 @@ export default {
             // добавляем ось Y
             let yAxis = svg.append("g")
               .attr("class",`yAxis-${i}`)
-              .call(d3.axisLeft(y[i]).tickValues(tickvals).tickFormat(x => `${x} ${metricUnits[i].units}`));
+              .call(d3.axisLeft(y[i]).tickValues(tickvals).tickFormat(x => `${x} ${metricUnits[i] ? metricUnits[i].units : ''}`));
 
             yAxis.selectAll(".tick>text").each(function(d) {
               let w = this.getBBox().width;
@@ -881,7 +874,7 @@ export default {
             // добавляем ось Y
             let yAxis = svg.append("g")
               .attr("class",`yAxis-${i}`)
-              .call(d3.axisLeft(y[i]).tickValues(tickvals).tickFormat(x => `${x} ${metricUnits[i].units}`));
+              .call(d3.axisLeft(y[i]).tickValues(tickvals).tickFormat(x => `${x} ${metricUnits[i] ? metricUnits[i].units : ''}`));
 
             yAxis.selectAll(".tick>text").each(function(d) {
               let w = this.getBBox().width;
@@ -895,9 +888,9 @@ export default {
 
           textsNodes.forEach( (item,l) => {
             if (textsNodes.length > 1) {
-              if (l == 0) {
+              if (l === 0) {
                 item.style["transform"] = "translateY(-5px)";
-              } else if (l == textsNodes.length-1) {
+              } else if (l === textsNodes.length-1) {
                 item.style["transform"] = "translateY(5px)";
               }
             } else {
@@ -912,14 +905,12 @@ export default {
           //     }
           //   })
 
-
-          if (Object.keys(metricOPt).length == 0 || metricOPt.type == 'Line chart') {
+          if (Object.keys(metricOPt).length === 0 || metricOPt.type === 'Line chart') {
 
             // строим основную линию
 
             line.push(svg.append('g')  // основная линия графика
               .attr("clip-path", `url(#clip-${that.id})`));
-
 
             let nullValue = -1;
             let dotDate = null;
@@ -935,7 +926,6 @@ export default {
             brushObj['startX'] = 0;
             brushObj['endX'] = 0;
 
-
             if (extraDot.length > 0) {
               extraDot.forEach( (item,j) => {
                 if (metric == item.column) {
@@ -949,22 +939,19 @@ export default {
                 return item
               }
             })
-            //console.log(cutData)
-
-
 
             if(nullValue == -1) {
 
               cutData.forEach( line => {
-                if (!Number(line[metric]) && line[metric] != 0) {
-                  onelinesWithBreak.length == 1 ? mustSee.push(onelinesWithBreak[0]) : false;
+                if (!Number(line[metric]) && line[metric] !== 0) {
+                  onelinesWithBreak.length === 1 ? mustSee.push(onelinesWithBreak[0]) : false;
                   linesWithBreak.push(onelinesWithBreak);
                   onelinesWithBreak = [];
                 } else {
                   onelinesWithBreak.push(line);
                 }
               })
-              onelinesWithBreak.length == 1 ? mustSee.push(onelinesWithBreak[0]) : false;
+              onelinesWithBreak.length === 1 ? mustSee.push(onelinesWithBreak[0]) : false;
               linesWithBreak.push(onelinesWithBreak);
               AllLinesWithBreak[i] = (linesWithBreak);
               linesWithBreak.forEach( (lineItself,j) => {
@@ -994,18 +981,11 @@ export default {
                     .attr("opacity", "0.3")
 
                 }
-
-
               })
-
-
               dotDate = cutData;
-
             } else {
               dotDate = [extraDot[nullValue]];
             }
-
-
 
             svg
               .append("g")
@@ -1054,7 +1034,7 @@ export default {
                 }
                 let text =  '';
                 that.metrics.forEach( key => {
-                  if (key == xMetric) {
+                  if (key === xMetric) {
                     text += `<p><span>${key}</span> : ${xVal}</p>`;
                   } else {
                     text += `<p><span>${key}</span> : ${d[key]}</p>`;
@@ -1083,7 +1063,7 @@ export default {
 
                 allDotHover = svg.selectAll('circle').nodes().filter( dot => {
                   if(dot.classList.contains('dot')){
-                    if (dot.getAttribute('cx') == x(d[xMetric]*secondTransf) && dot['__data__'][dot.getAttribute('metric')] != null ) {
+                    if (dot.getAttribute('cx') === x(d[xMetric]*secondTransf) && dot['__data__'][dot.getAttribute('metric')] != null ) {
                       dot.style="opacity:1";
                       return dot
                     }
@@ -1101,12 +1081,12 @@ export default {
               .on("mouseout", function(d) {
                 let opacity = 1;
                 if (!this.getAttribute("data-last-dote") ) {
-                  if (nullValue == -1) {
+                  if (nullValue === -1) {
                     opacity = 0;
                   }
                   allDotHover.forEach( dot => {
 
-                    if (extraDot.indexOf(dot['__data__']) == -1) {
+                    if (extraDot.indexOf(dot['__data__']) === -1) {
                       dot.style=`opacity:0`;
                     }
                     if (dot.getAttribute("data-with-caption")) {
@@ -1122,7 +1102,7 @@ export default {
 
 
                 mustSee.forEach( item => {
-                  if (item[metric] == d[metric]) {
+                  if (item[metric] === d[metric]) {
                     opacity =  1;
                   }
                 })
@@ -1217,7 +1197,7 @@ export default {
 
             brushObj['selectionUp'] = () => {
               brushObj.mouseDown = false;
-              if (brushObj.direction == 'left') {
+              if (brushObj.direction === 'left') {
                 let change = 0;
                 change = brushObj.startX;
                 brushObj.startX = brushObj.endX;
@@ -1231,7 +1211,7 @@ export default {
 
             brushObj['clearBrush'] = () => {
               brushObj.selections = brush.selectAll(`.selection-${i}`).nodes();
-              if (brushObj.selections.length != 0) {
+              if (brushObj.selections.length !== 0) {
                 brushObj.selections.forEach( (item,i) => {
                   brushObj.selections[i].remove()
                 })
@@ -1239,7 +1219,7 @@ export default {
             }
 
 
-          } else if (Object.keys(metricOPt).length != 0 || metricOPt.type == 'Bar chart') {
+          } else if (Object.keys(metricOPt).length !== 0 || metricOPt.type === 'Bar chart') {
 
             //line.push('Bar chart');  // добовляем в массив заглушку, чтобы собюсти порядок следования линий для линейных графиков
             let allDotHover = [];
@@ -1309,14 +1289,14 @@ export default {
               })
               .attr("height", function(d,j) {
                 if (lastDot) {
-                  if (j == cutData.length - 1) {
+                  if (j === cutData.length - 1) {
                     putLabelDot('data-last-bar','last-bar-text',d,y[i](d[metricOPt.name])-5,metricOPt.name,this,that,'bar');
                   }
                 }
                 if (d[`_${metricOPt.name}_caption`]) {
                   putLabelDot('data-with-caption','caption-bar-text',d,y[i](d[metricOPt.name])-5,`_${metricOPt.name}_caption`,this,that,'bar');
                 }
-                if (d[metricOPt.name] == 0) {
+                if (d[metricOPt.name] === 0) {
                   return 0
                 }
                 if (negative) {
@@ -1336,7 +1316,7 @@ export default {
                 }
                 let text =  '';
                 that.metrics.forEach( key => {
-                  if (key == xMetric) {
+                  if (key === xMetric) {
                     text += `<p><span>${key}</span> : ${xVal}</p>`;
                   } else {
                     text += `<p><span>${key}</span> : ${d[key]}</p>`;
@@ -1366,7 +1346,7 @@ export default {
 
                 allDotHover = svg.selectAll('circle').nodes().filter( dot => {
                   if(dot.classList.contains('dot')){
-                    if (dot['__data__'][xMetric] == d[xMetric]*secondTransf && dot['__data__'][dot.getAttribute('metric')] != null ) {
+                    if (dot['__data__'][xMetric] === d[xMetric]*secondTransf && dot['__data__'][dot.getAttribute('metric')] != null ) {
                       dot.style="opacity:1";
                       return dot
                     }
@@ -1378,7 +1358,7 @@ export default {
               .on("mouseout", function() {
                 if (!this.getAttribute("data-last-bar") ) {
                   allDotHover.forEach( dot => {
-                    if (extraDot.indexOf(dot['__data__']) == -1) {
+                    if (extraDot.indexOf(dot['__data__']) === -1) {
                       dot.style=`opacity:0`;
                     }
                     if (dot.getAttribute("data-with-caption")) {
@@ -1417,7 +1397,7 @@ export default {
       function verticalLineX() {
 
         let linesX = svg.selectAll(`.grid-line-x`).nodes();
-        if (linesX.length != 0) {
+        if (linesX.length !== 0) {
           linesX.forEach( (item,i) => {
             linesX[i].remove()
           })
@@ -1493,7 +1473,7 @@ export default {
           .style('fill', that.theme.$main_text)
           .text(d[metricText])
 
-        if (elem == 'line') {
+        if (elem === 'line') {
           text
             .on("mouseover", function() {
               if(brushObj.mouseDown) {
@@ -1546,14 +1526,13 @@ export default {
         //   line.select(`.brush`).call(brush.move, null);  // убираем область выделения
         // }
         brushObj.clearBrush();
-
         if (time) {
           xAxis.transition().duration(secondTransf)
             .call(d3.axisBottom(x)
-              .tickFormat(d3.timeFormat('%d-%m-%Y '))
+              .tickFormat(d3.timeFormat(timeFormat))
               .tickValues(
                 x.ticks().filter( (item,i) => {
-                  if (i%2 == 0) {
+                  if (i%2 === 0) {
                     return item
                   }
                 })
@@ -1566,7 +1545,7 @@ export default {
 
         verticalLineX();
 
-        if (id == -1) {
+        if (id === -1) {
           metricsName.forEach( (item,i) => {
 
             changeZoom(1000,i);   // вызываем функцию которая перересует все линии и точки как надо
@@ -1577,15 +1556,14 @@ export default {
 
 
         svg.on("dblclick",function(){  // если дважды щелкнуть по любому месту
-
           if (time) {
             x.domain(d3.extent(data, function(d) {   return  new Date(d[xMetric]*secondTransf) })); // то вернем ось х в исходное состояние
             xAxis.transition().duration(secondTransf)
               .call(d3.axisBottom(x)
-                .tickFormat(d3.timeFormat('%d-%m-%Y '))
+                .tickFormat(d3.timeFormat(timeFormat))
                 .tickValues(
                   x.ticks().filter( (item,i) => {
-                    if (i%2 == 0) {
+                    if (i%2 === 0) {
                       return item
                     }
                   })
@@ -1602,18 +1580,16 @@ export default {
           // metricsName.forEach( (item,i) => {
           //   changeZoom(300,i);    // вызываем функцию которая перересует все линии и точки как надо
           // })
-          if (id == -1) {
-            metricsName.forEach( (item,i) => {
-              changeZoom(300,i);   // вызываем функцию которая перересует все линии и точки как надо
-            })
-          } else {
-            changeZoom(300,id);   // вызываем функцию которая перересует все линии и точки как надо
-          }
+          // if (id ===-1) {
+          metricsName.forEach( (item,i) => {
+            changeZoom(300,i);   // вызываем функцию которая перересует все линии и точки как надо
+          })
+          // } else {
+          //   changeZoom(300,id);   // вызываем функцию которая перересует все линии и точки как надо
+          // }
           // changeZoom(300);  // и вернем все линии  в исходное состояние
 
         });
-
-
 
         function changeZoom(dauration,i) {  // функция которая перерисовывает все линии
           let lineChange;
@@ -1696,9 +1672,9 @@ export default {
         let count = {};
         Object.values(data).forEach( item => {
           Object.keys(item).forEach( (key,i) => {
-            if (i !=0) {
+            if (i !== 0) {
 
-              if (!Number(item[key]) && item[key] != 0) {
+              if (!Number(item[key]) && item[key] !== 0) {
                 if (!nullArr.includes(key)) {
                   nullArr.push(key)
                 }
@@ -1715,7 +1691,7 @@ export default {
 
 
         nullArr = nullArr.filter( item => {
-          if (count[item] == 1) {
+          if (count[item] === 1) {
             return item
           }
         })
@@ -1725,7 +1701,7 @@ export default {
 
             Object.values(data).filter( item => {
 
-              if (Number(item[key]) || Number(item[key]) == 0) {
+              if (Number(item[key]) || Number(item[key]) === 0) {
                 value.push({...item,...{column: key}});
               }
             });
@@ -1764,26 +1740,26 @@ export default {
           action: tockens[i].action,
           capture: tockens[i].capture,
         }
-        if (tockens[i].elem == this.id && tockens[i].action == action && tockens[i].capture == 'pointX') {
+        if (tockens[i].elem === this.id && tockens[i].action === action && tockens[i].capture === 'pointX') {
           setTocken(point.x);
 
-        } else if (tockens[i].elem == this.id && tockens[i].action == action && tockens[i].capture == 'pointY') {
+        } else if (tockens[i].elem === this.id && tockens[i].action === action && tockens[i].capture === 'pointY') {
           setTocken(point.y);
-        } else if (tockens[i].elem == this.id && tockens[i].action == action && tockens[i].capture == 'start') {
+        } else if (tockens[i].elem === this.id && tockens[i].action === action && tockens[i].capture === 'start') {
           setTocken(point[0]);
-        } else if (tockens[i].elem == this.id && tockens[i].action == action && tockens[i].capture == 'end') {
+        } else if (tockens[i].elem === this.id && tockens[i].action === action && tockens[i].capture === 'end') {
           setTocken(point[1]);
         }
       })
       
       let events = this.$store.getters.getEvents({idDash: this.idDash, event: 'onclick', element: this.id, partelement: 'point'});
 
-      if (events.length != 0) {
+      if (events.length !== 0) {
         events.forEach( item => {
-          if(item.action == 'set'){
+          if(item.action === 'set'){
             this.$store.commit('letEventSet', {events: events, idDash: this.idDash,  });
-          } else if (item.action == 'go') {
-            if (action != 'select') {
+          } else if (item.action === 'go') {
+            if (action !== 'select') {
               this.$store.commit('letEventGo', {event: item, idDash: this.idDash, route: this.$router, store: this.$store });
              // this.$router.push(`/dashboards/${item.target.toLowerCase()}`);
             }