+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - hover on first row in table
+- time format on x-axis while zooming
 
 ## [1.63.0]